// packages/builtin-tool-fi-agent/src/types.ts
export const FiAgentApiName = {
  analyzeVisualMedia: 'analyzeVisualMedia',
  callSubAgent: 'callSubAgent',
  callSubAgents: 'callSubAgents',
  clearTodos: 'clearTodos',
  createPlan: 'createPlan',
  createTodos: 'createTodos',
  updatePlan: 'updatePlan',
  updateTodos: 'updateTodos',
} as const;